requirements_dev.txt
bgen_reader
cyvcf2
equinox
glimix_core
jax
jaxlib
pandas>=1.5.0
pandas-plink==2.2.9
pre-commit
pymdown-extensions
pytest-accept
pytest-cov
scipy
setuptools
sphinx>=5
sphinx_immaterial
